* Paket:    - control: bittorrent zu Depends hinzugefügt.    - rules: Code zum Herunterladen und Kompilieren von ctorrent hinzugefügt.    - postinst:      o update-rc.d für linbo-bittorrent hinzugefügt.      o in $LINBODIR/start.conf wird die Server-IP geprüft.      o in /etc/default/bittorrent werden die Optionen PORT, DAEMONUSER, REANNOUNCE_INTERVAL und ALLOWED_DIR geprüft und ggf. gesetzt.      o Benutzer bittorrent wird erstellt, falls nicht vorhanden.      o in den start.conf-Dateien unter $LINBODIR wird die obsolete Option UseMulticast nach DownloadType geändert.    - neue Dateien:      o debian/linbo-bittorrent.init -> /etc/init.d/linbo-bittorrent.      o debian/linbo-bittorrent.default -> /etc/default/linbo-bittorrent.   * Serverseitige Skripte:     - init-Skript für linbo-bittorrent hinzugefügt (s.o.).     - default-Konfigurationsdatei für linbo-bittorrent hinzugefügt (s.o.).     - rsync-post.sh:       o Multicast- und Bittorrent-Dienste werden nach dem Hochladen von Image-Dateien neu gestartet.       o mehrfacher Eintrag einer schon in der Workstationsdatei vorhandenen Zeile wird nun vermieden.     - start.conf: Neue globale Option DownloadType, ersetzt UseMulticast, kann die Werte rsync|torrent|multicast enthalten, Standardwert ist rsync (s.o.).     - Beispieldateien: UseMulticast durch DownloadType ersetzt (s.o.).     - helperfunctions.sh: neue Funktionsbibliothek für LINBO-Serverskripte.  * Linbofs:    - ctorrent-Binary hinzugefügt.    - linbo_cmd:      o Torrent-Funktionalität ergänzt.      o /cache bleibt nun während der gesamten LINBO-Session gemountet.      o ip(): Konnte keine Client-IP ermittelt werden, wird "OFFLINE" zurückgegeben.      o initcache(): Obsolete Image- und Linbofs-Dateien werden entfernt.      o initcache(): Benutzt nun download_if_newer() um Cachedateien zu aktualisieren. Damit lädt Multicast Images nicht mehr herunter, obwohl sie schon komplett im Cache liegen.      o initcache(): Von der Gui übergebener DownloadType wird ignoriert, stattdessen wird der DownloadType benutzt, der in der start.conf definiert ist.      o downloadtype(): Neue Funktion, ermittelt den in der start.conf gesetzten DownloadType.      o localmode(): Netzwerkerkennung vereinfacht, benutzt nun das Flag /tmp/network.ok, das von init.sh gesetzt wird.      o download_if_newer(): multicast.list wird nun von init.sh heruntergeladen.    - init.sh:      o network(): Wenn start.conf vom Server heruntergeladen werden konnte, wird das Flag /tmp/network.ok gesetzt (s.o.).      o network(): torrent-client.conf und multicast.list werden je nach DownloadType nach / heruntergeladen. So werden sie pro LINBO-Session nur einmal heruntergeladen anstatt von linbo_cmd bei jedem Aufruf von download_if_newer().      o init_setup(): Server-IP wird nun aus /proc/cmdline geparst, falls sie mit server=x.x.x.x gesetzt wurde. Falls nicht gesetzt wird die per DHCP ermittelte Server-IP verwendet.      o get_server(): Server-IP aus /tmp/dhcp.log wird nun aus der Variable siaddr ermittelt, scheint zuverlässiger als serverid.

+++ linbo/init.sh
@@ -70,9 +70,12 @@ init_setup(){
  mount -t proc /proc /proc
  echo 0 >/proc/sys/kernel/printk
 
+ # parse kernel cmdline
  CMDLINE="$(cat /proc/cmdline)"
- echo "$CMDLINE" | grep -q debug && debug=yes
- echo "$CMDLINE" | grep -q useide && useide=yes
+ case "$CMDLINE" in *\ useide*) useide=yes;; esac
+ case "$CMDLINE" in *\ debug*) debug=yes;; esac
+ case "$CMDLINE" in *\ nonetwork*|*\ localmode*) localmode=yes;; esac
+ for i in $CMDLINE; do case "$i" in *=*) eval "$i";; esac; done
 
  mount -t sysfs /sys /sys
  mount -n -o mode=0755 -t tmpfs tmpfs /dev
@@ -170,17 +173,17 @@ $(nslookup "$1" 2>/dev/null)
  return 1
 }
 
-# Get Gateway address as server.
+# Get server address.
 get_server(){
  local ip=""
  local a=""
  local b=""
- # Try servername from dhcp.log first.
+ # First try servername from dhcp.log:siaddr.
  if [ -f "/tmp/dhcp.log" ]; then
-  ip="`grep ^serverid /tmp/dhcp.log | tail -1 | cut -f2 -d"'"`"
+  ip="`grep ^siaddr /tmp/dhcp.log | tail -1 | cut -f2 -d"'"`"
   [ -n "$ip" ] && { echo "$ip"; return 0; }
  fi
- # Then guess from route
+ # Second guess from route.
  while read a b relax; do
   case "$a" in 0.0.0.0)
    ip="$b"
@@ -193,12 +196,22 @@ $(route -n)
  return 1
 }
 
+# get DownloadType from start.conf
+downloadtype(){
+ local RET=""
+ if [ -s /start.conf ]; then
+  RET="$(grep -i ^downloadtype /start.conf | tail -1 | awk -F= '{ print $2 }' | awk '{ print $1 }' | tr A-Z a-z)"
+  # get old option for compatibility issue
+  if [ -z "$RET" ]; then
+   RET="$(grep -i ^usemulticast /start.conf | tail -1 | awk -F= '{ print $2 }' | awk '{ print $1 }' | tr A-Z a-z)"
+   [ "$RET" = "yes" ] && RET="multicast"
+  fi
+ fi
+ echo "$RET"
+}
+
 network(){
- case "$(cat /proc/cmdline)" in *\ nonetwork|*\ localmode*)
-  touch /tmp/linbo-network.done
-  return 0
-  ;;
- esac
+ [ -n "$localmode" ] && { touch /tmp/linbo-network.done; return 0; }
  rm -f /tmp/linbo-network.done
  if [ -n "$ipaddr" ]; then
   [ -n "$netmask" ] && nm="netmask $netmask" || nm=""
@@ -221,9 +234,16 @@ network(){
  # Move away old start.conf and look for updates
  mv -f start.conf start.conf.dist
  if [ -n "$server" ]; then
+  echo "linbo_server='$server'" >> /tmp/dhcp.log
   for i in "start.conf-$ipaddr" "start.conf"; do
    rsync -L "$server::linbo/$i" "start.conf" >/dev/null 2>&1 && break
   done
+  # also look for other needed files
+  case "$(downloadtype)" in
+   torrent) dlfile="torrent-client.conf" ;;
+   multicast) dlfile="multicast.list" ;;
+  esac
+  [ -n "$dlfile" ] && rsync -L "$server::linbo/$dlfile" "$dlfile" >/dev/null 2>&1
  fi
  if [ ! -s start.conf ]; then
   # No new version / no network available, look for a cached copy.
@@ -231,6 +251,9 @@ network(){
    sleep 1
   done
   copyfromcache start.conf
+ else
+  # flag for network connection
+  echo > /tmp/network.ok
  fi
  # Still nothing new, revert to old version.
  [ -s start.conf ] || mv -f start.conf.dist start.conf
